+++ tide/core/types.h
@@ -202,6 +202,22 @@ inline std::ostream& operator<<(std::ostream& str, const QPointF& p)
     return str;
 }
 
+inline std::ostream& operator<<(std::ostream& str, const ScreenState state)
+{
+    switch (state)
+    {
+    case ScreenState::ON:
+        str << "ON";
+        return str;
+    case ScreenState::OFF:
+        str << "OFF";
+        return str;
+    default:
+        str << "UNDEF";
+    }
+    return str;
+}
+
 inline std::ostream& operator<<(std::ostream& str, const QRectF& r)
 {
     str << r.x() << ',' << r.y() << ' ' << r.width() << 'x' << r.height();